requirements.txt
django
pylint
psycopg2-binary
requests
python-dotenv
django-debug-toolbar